Reverted db key mapping, moved export files
