+++ mappings/export/exportTranslation.ts
@@ -1,0 +1,49 @@
+/*
+  Export for FE translation files
+  Output filename/directory - /mappings/translations/model_plan_translations.json
+  For BE use for storing translated data to database
+*/
+import * as fs from 'fs';
+
+import basics from '../../src/i18n/en-US/modelPlan/basics';
+import beneficiaries from '../../src/i18n/en-US/modelPlan/beneficiaries';
+import collaborators from '../../src/i18n/en-US/modelPlan/collaborators';
+import generalCharacteristics from '../../src/i18n/en-US/modelPlan/generalCharacteristics';
+import modelPlan from '../../src/i18n/en-US/modelPlan/modelPlan';
+import opsEvalAndLearning from '../../src/i18n/en-US/modelPlan/opsEvalAndLearning';
+import participantsAndProviders from '../../src/i18n/en-US/modelPlan/participantsAndProviders';
+import payments from '../../src/i18n/en-US/modelPlan/payments';
+import { getKeys } from '../../src/types/translation';
+
+import { processDataMapping } from './util';
+
+export const translationSections = {
+  model_plan: modelPlan,
+  plan_basics: basics,
+  plan_general_characteristics: generalCharacteristics,
+  plan_participants_and_providers: participantsAndProviders,
+  plan_beneficiaries: beneficiaries,
+  plan_ops_eval_and_learning: opsEvalAndLearning,
+  plan_payments: payments,
+  plan_collaborators: collaborators
+};
+
+export const parseTypscriptToJSON = (translations: any, outputFile: string) => {
+  const jsonString = JSON.stringify(translations, null, 2);
+
+  fs.writeFileSync(outputFile, jsonString);
+};
+
+(function main() {
+  const transformedTranslationSections = processDataMapping(
+    translationSections
+  );
+
+  // Create JSON file for each translation task list section
+  getKeys(transformedTranslationSections).forEach(section =>
+    parseTypscriptToJSON(
+      transformedTranslationSections[section],
+      `./mappings/translation/${String(section)}.json`
+    )
+  );
+})();
